chore: rename package manager
