+++ internal/package_handler/package_test.go
@@ -1,4 +1,4 @@
-package package_manager
+package package_handler
 
 import (
 	"os"
@@ -84,8 +84,8 @@ func TestCheck(t *testing.T) {
 
 	for _, tc := range ts {
 		t.Run(tc.name, func(t *testing.T) {
-			pkgManager := NewPackageHandler(tc.pkgFolder)
-			err := pkgManager.Check()
+			pkgHandler := NewPackageHandler(tc.pkgFolder)
+			err := pkgHandler.Check()
 			assert.ErrorIs(t, err, tc.err)
 		})
 	}
