+++ internal/package_handler/testdata/mock-avs/checksum.txt
