+++ internal/package_handler/testdata/mock-avs/pkg/manifest.yml
