+++ internal/package_handler/testdata/mock-avs/pkg/sepolia/docker-compose.yml
